requirements.txt
google-generativeai>=0.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
deepdiff>=7.0.0
rich>=13.0.0
